Delete README.md
In this project we are given a Diabetes Data set consisting of following features -

['Pregnancies', 'Glucose', 'BloodPressure', 'SkinThickness', 'Insulin', 'BMI', 'DiabetesPedigreeFunction', 'Age', 'Outcome']

We have to predict whether a person is suffering from diabetes or not (Binary Classification). 
Classification Task - Classify a person as 0 or 1 (Diabetic or Not) using K-Nearest Neighbors classifier.
We are saving our predictions of test cases in a CSV file. File should have same structure as that of Sample Submission File.